registration

+++ app/src/main/java/com/hollybits/socialpetnetwork/models/Pet.java
@@ -10,14 +10,14 @@ public class Pet {
     private Long age;
     private Sex sex;
     private String tagNumber;
-    private Long weight;
+    private Double weight;
     private Attitude attitude;
 
     public Pet(){
 
     }
 
-    public Pet(String name, Breed breed, Long age, Sex sex, String tagNumber, Long weight, Attitude attitude) {
+    public Pet(String name, Breed breed, Long age, Sex sex, String tagNumber, Double weight, Attitude attitude) {
         this.name = name;
         this.breed = breed;
         this.age = age;
@@ -67,11 +67,11 @@ public class Pet {
         this.tagNumber = tagNumber;
     }
 
-    public Long getWeight() {
+    public Double getWeight() {
         return weight;
     }
 
-    public void setWeight(Long weight) {
+    public void setWeight(Double weight) {
         this.weight = weight;
     }
